fix(csa-scheduler): treat gemini API_KEY_INVALID 400 as failover-eligible auth error (#1848)
A mid-run gemini-cli provider-400 carrying an OAuth/key-exhaustion signal
(`reason: "API_KEY_INVALID"` / "API key not valid. Please pass a valid API
key.") was classified as a generic `HTTP 400`, which `detect_rate_limit`
gates behind the 30s init-failure window. When the 400 was raised after that
window (e.g. a plan/mktd RECON dimension failing at ~39s, once earlier
dimensions had already succeeded), the classifier returned no failover, so
the tier step aborted via `on_fail=abort` and discarded prior RECON work --
instead of advancing to the next tier candidate the way `csa review` already
did in the same OAuth-dead window.

Add two gemini-cli failover patterns -- `api_key_invalid` (matches the
structured Google `reason` code) and `api key not valid` (matches the human
message) -- classed as `auth_unavailable`. That reason bypasses the
init-failure window, so a key/OAuth-exhaustion 400 stays failover-eligible
mid-run, not only at startup. Both carry `quota_exhausted = false`: an auth
rejection is not permanent quota exhaustion, so it never trips the permanent
self-kill path.

Anti-over-broadening guard (#1736): eligibility keys on the specific
OAuth/key-exhaustion marker, never "any 400". A generic non-auth 400 past the
init window still classifies as `HTTP 400` and surfaces as an error rather
than being silently masked by failover. `detect_rate_limit` is the single
classifier shared by all three paths (`csa run`, plan/mktd RECON, and
`csa review`), so this one fix repairs them together.

Document the within-tier routing precedence in docs/configuration.md, and
record that `[preferences].tool_priority` reordering of tier candidates is a
soft preference (never a whitelist) and is intentionally left unchanged.

To keep tier_model_fallback.rs under the 8000-token monolith budget once the
new cases are added, its `#[cfg(test)] mod tests` is extracted verbatim to a
sibling `tier_model_fallback_tests.rs` wired via `#[path]` (the same idiom as
failover_trace.rs); `*_tests.rs` files are exempt from the guard. No test
logic changed in the move.

Tests cover both directions: csa-scheduler rate_limit (API_KEY_INVALID JSON
-> auth_unavailable, init window bypassed; message-only phrasing ->
auth_unavailable; generic 400 -> stays HTTP 400, init-gated) and
cli-sub-agent tier_model_fallback (mid-run API_KEY_INVALID advances at 39s;
generic 400 does NOT advance at 39s; generic 400 still advances at 5s).
