+++ Cargo.toml
@@ -20,7 +20,7 @@ default-members = [
 resolver = "2"
 
 [workspace.package]
-version = "0.1.874"
+version = "0.1.875"
 edition = "2024"
 rust-version = "1.88"
 license = "Apache-2.0"
