+++ crates/cli-sub-agent/src/tier_model_fallback.rs
@@ -464,205 +464,5 @@ pub(crate) fn persist_result_warning(project_root: &Path, session_id: &str, warn
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        TierAttemptFailure, earliest_backend_reset_window,
-        format_all_models_failed_reason_with_reset, opaque_total_exhaustion_message,
-        ordered_tier_candidates, parse_backend_reset_duration,
-    };
-    use csa_config::{GlobalConfig, ProjectConfig, ToolConfig};
-    use csa_core::types::ToolName;
-    use std::collections::HashMap;
-
-    fn project_config_with_tier(
-        tier_name: &str,
-        models: &[&str],
-        enabled_tools: &[&str],
-    ) -> ProjectConfig {
-        let mut tool_map = HashMap::new();
-        for tool in csa_config::global::all_known_tools() {
-            let name = tool.as_str();
-            tool_map.insert(
-                name.to_string(),
-                ToolConfig {
-                    enabled: enabled_tools.contains(&name),
-                    ..Default::default()
-                },
-            );
-        }
-
-        let mut cfg = ProjectConfig {
-            schema_version: csa_config::config::CURRENT_SCHEMA_VERSION,
-            project: csa_config::ProjectMeta {
-                name: "test".to_string(),
-                created_at: chrono::Utc::now(),
-                max_recursion_depth: 5,
-            },
-            resources: csa_config::ResourcesConfig::default(),
-            acp: Default::default(),
-            tools: tool_map,
-            review: None,
-            debate: None,
-            tiers: HashMap::new(),
-            tier_mapping: HashMap::new(),
-            aliases: HashMap::new(),
-            tool_aliases: HashMap::new(),
-            preferences: None,
-            github: None,
-            session: Default::default(),
-            memory: Default::default(),
-            hooks: Default::default(),
-            run: Default::default(),
-            execution: Default::default(),
-            session_wait: None,
-            preflight: Default::default(),
-            vcs: Default::default(),
-            filesystem_sandbox: Default::default(),
-        };
-        cfg.tiers.insert(
-            tier_name.to_string(),
-            csa_config::config::TierConfig {
-                description: "Test tier".to_string(),
-                models: models.iter().map(|spec| (*spec).to_string()).collect(),
-                strategy: csa_config::TierStrategy::default(),
-                token_budget: None,
-                max_turns: None,
-            },
-        );
-        cfg
-    }
-
-    #[test]
-    fn tier_fallback_prefers_original_tool_then_keeps_full_tier() {
-        let _availability = crate::test_env_lock::ScopedEnvVarRestore::set(
-            crate::run_helpers::TEST_ASSUME_TOOLS_AVAILABLE_ENV,
-            "1",
-        );
-        let cfg = project_config_with_tier(
-            "quality",
-            &[
-                "codex/openai/gpt-5.4/high",
-                "gemini-cli/google/gemini-3.1-pro-preview/xhigh",
-                "claude-code/anthropic/sonnet-4.6/xhigh",
-            ],
-            &["codex", "gemini-cli", "claude-code"],
-        );
-
-        let candidates = ordered_tier_candidates(
-            ToolName::Codex,
-            Some("codex/openai/gpt-5.4/high"),
-            Some("quality"),
-            Some(&cfg),
-            None,
-            true,
-            &["codex".to_string()],
-        );
-
-        assert_eq!(
-            candidates,
-            vec![
-                (
-                    ToolName::Codex,
-                    Some("codex/openai/gpt-5.4/high".to_string()),
-                ),
-                (
-                    ToolName::GeminiCli,
-                    Some("gemini-cli/google/gemini-3.1-pro-preview/xhigh".to_string()),
-                ),
-                (
-                    ToolName::ClaudeCode,
-                    Some("claude-code/anthropic/sonnet-4.6/xhigh".to_string()),
-                ),
-            ]
-        );
-    }
-
-    #[test]
-    fn no_tier_fallback_uses_global_tool_priority() {
-        let _availability = crate::test_env_lock::ScopedEnvVarRestore::set(
-            crate::run_helpers::TEST_ASSUME_TOOLS_AVAILABLE_ENV,
-            "1",
-        );
-        let mut global = GlobalConfig::default();
-        global.preferences.tool_priority =
-            vec!["claude-code".to_string(), "gemini-cli".to_string()];
-        let candidates =
-            ordered_tier_candidates(ToolName::Codex, None, None, None, Some(&global), true, &[]);
-
-        assert!(candidates.starts_with(&[
-            (ToolName::Codex, None),
-            (ToolName::ClaudeCode, None),
-            (ToolName::GeminiCli, None),
-        ]));
-    }
-
-    #[test]
-    fn opaque_total_exhaustion_uses_provider_reset_window() {
-        let failure_reason = "all tier-4-critical models failed: \
-            gemini-cli/google/gemini-3.1-pro-preview/xhigh=auth_unavailable; quota will reset after 14h, \
-            codex/openai/gpt-5.5/xhigh=HTTP 429; reset in 1h";
-
-        assert_eq!(
-            opaque_total_exhaustion_message(
-                Some("auth_unavailable; HTTP 429"),
-                Some(failure_reason)
-            )
-            .as_deref(),
-            Some(
-                "review unavailable: all tier-4-critical backends rate-limited; earliest reset ~1h 0m"
-            )
-        );
-    }
-
-    #[test]
-    fn opaque_total_exhaustion_omits_unparseable_reset_window() {
-        let failure_reason = "all tier-4-critical models failed: \
-            gemini-cli/google/gemini-3.1-pro-preview/xhigh=auth_unavailable; reset pending, \
-            codex/openai/gpt-5.5/xhigh=HTTP 429; reset unknown";
-
-        assert_eq!(
-            opaque_total_exhaustion_message(
-                Some("auth_unavailable; HTTP 429"),
-                Some(failure_reason)
-            )
-            .as_deref(),
-            Some("review unavailable: all tier-4-critical backends rate-limited")
-        );
-    }
-
-    #[test]
-    fn all_models_failed_reason_uses_earliest_backend_reset_window() {
-        let reset_windows = [
-            parse_backend_reset_duration("provider said quota will reset after 14h"),
-            parse_backend_reset_duration("provider said reset in 1h"),
-            parse_backend_reset_duration("provider said reset unknown"),
-        ];
-        let parseable_reset_windows = reset_windows.into_iter().flatten().collect::<Vec<_>>();
-        let failures = vec![
-            TierAttemptFailure {
-                model_spec: "gemini-cli/google/gemini-3.1-pro-preview/xhigh".to_string(),
-                reason: "auth_unavailable".to_string(),
-                quota_exhausted: Some(false),
-            },
-            TierAttemptFailure {
-                model_spec: "codex/openai/gpt-5.5/xhigh".to_string(),
-                reason: "HTTP 429".to_string(),
-                quota_exhausted: Some(false),
-            },
-        ];
-
-        assert_eq!(
-            format_all_models_failed_reason_with_reset(
-                Some("tier-4-critical"),
-                &failures,
-                earliest_backend_reset_window(&parseable_reset_windows),
-            )
-            .as_deref(),
-            Some(
-                "all tier-4-critical models failed: \
-gemini-cli/google/gemini-3.1-pro-preview/xhigh=auth_unavailable, \
-codex/openai/gpt-5.5/xhigh=HTTP 429; earliest_reset=1h 0m"
-            )
-        );
-    }
-}
+#[path = "tier_model_fallback_tests.rs"]
+mod tests;
